document_extractor/pyproject.toml
[project]
name = "document-extractor"
version = "0.2.0"
description = "Document Extractor Agent using Docling (Parse) + Qwen3.5 (Extract) via ADK"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-adk>=1.5.0",
    "google-genai>=1.14.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "docling>=2.0.0",
    "PyPDF2>=3.0.0",
    "python-dotenv>=1.0.0",
    "pytest>=7.0.0",
]

[project.optional-dependencies]
# Keep pymupdf as optional fallback for legacy code
legacy = [
    "pymupdf>=1.24.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.pytest.ini_options]
markers = [
    "e2e: end-to-end tests requiring external services (Ollama, Docling)",
]